refactor(bonuses): finish bonus selection cutover
